requirements.txt
argparse
requests
polygon-api-client
dotenv